Keep track of package-lock.json for security reason.
In order to permit the security audit of installed dependencies,
we need to track the exact installed package version inside git.

A good side effect is that the same version will be installed
everywhere.

Needed for the security concepts compliance.

The file is currently not added in .gitignore. But if that causes more
troubles than the issue it should solve, I will add it later on.

Should be merged after:
https://github.com/LucasBrazi06/ev-dashboard-new/pull/480

Signed-off-by: Jerome Benoit <jerome.benoit@sap.com>